Done Attendance

+++ internal/domain/models/OfficeLocation.go
@@ -3,5 +3,5 @@ package models
 type OfficeLocationInfo struct {
 	Latitude  float64 `json:"latitude"`
 	Longitude float64 `json:"longitude"`
-	Radius    int     `json:"radius"`
+	Radius    float64 `json:"radius"`
 }
